move simulation log code to log.go
